requirements.txt
Flask==1.0
numpy==1.18.1
pickleshare==0.7.5
Pillow==7.0.0
pylint==2.4.4
scipy==1.4.1
tensorflow==1.15
tensorboard==1.15.0
tensorflow-hub==0.7.0
